merage 12,14,15,16,17,18,19,20,21,22
1.Fix code issue in original code;
2.Update UT/MT source code;
3.Add test code for stability;
4.open alg.DH funtion;
5.Add file lock to backup files;

+++ frameworks/huks_standard/main/crypto_engine/openssl/src/hks_openssl_dh.c
@@ -192,7 +192,7 @@ int32_t HksOpensslDhAgreeKey(const struct HksBlob *nativeKey, const struct HksBl
         return HKS_ERROR_CRYPTO_ENGINE_ERROR;
     }
 
-    if (HKS_KEY_BYTES(spec->keyLen) > DH_size(dh)) {
+    if (HKS_KEY_BYTES(spec->keyLen) > (uint32_t)DH_size(dh)) {
         ret = HKS_ERROR_INVALID_KEY_SIZE;
     } else {
         (void)memcpy_s(sharedKey->data, sharedKey->size, computeKey, HKS_KEY_BYTES(spec->keyLen));
